Fix ambiguous operator invocation
On more recent versions of C++, this line throws an ambiguous operator error since it can't determine if it should be using the CUtlSymbol equality operator or the built-in int equality operator.

This patch removes the ambiguity by explicitly stating that the CUtlSymbol equality operator should be used. This doesn't change the underlying behavior, it just fixes the ambiguous resolution error.